+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release notes
 
+## v2.6.1 (2025-06-10)
+- Improve illegal token detection
+- The body of @each statement now can be empty
+
 ## v2.6.0 (2025-06-07)
 - Improve parser implementation by making it more resilient to syntax errors. It now produces an IllegalNode ast token. I'll be able to use it to detect illegal parts in a file and show it with a red squiggly line
 - Improve testing for the parser by adding more checks